requirements_test.txt
pyjwt
requests
argparse
PyBrowserID==0.6
mock
